Revert "Fe interface"
