+++ frontend/package.json
@@ -17,8 +17,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router": "^6.24.1",
-    "react-router-dom": "^6.24.1",
-    "react-select": "^5.8.0"
+    "react-router-dom": "^6.24.1"
   },
   "devDependencies": {
     "@types/d3-graphviz": "^2.6.10",
