--- frontend/src/components/output/LLVMIR/LLVMIR.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import './llvmir.module.css';
-
-function LLVMIR() {
-
-  return (
-    <>
-      <div>LLVMIR</div>
-    </>
-  )
-}
-
-export default LLVMIR
