--- frontend/src/components/output/codeGPT/CodeGPT.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import './codeGPT.module.css';
-
-function CodeGPT() {
-
-  return (
-    <>
-      <div>Code GPT</div>
-    </>
-  )
-}
-
-export default CodeGPT
