--- frontend/src/components/output/outputMenuBar/outputMenuBar.module.css
@@ -1,22 +1,0 @@
-.navBar {
-  background-color: #f8f9fa;
-  padding: 10px;
-}
-
-.navList {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  justify-content: space-around;
-}
-
-.navItem {
-  cursor: pointer;
-  padding: 10px 20px;
-  transition: background-color 0.3s ease;
-}
-
-.navItem:hover {
-  background-color: #e9ecef;
-}
