--- frontend/src/components/output/terminalOutput/TerminalOutput.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import './terminalOutput.module.css';
-
-function TerminalOutput() {
-
-  return (
-    <>
-      <div>Terminal Output</div>
-    </>
-  )
-}
-
-export default TerminalOutput
