--- frontend/src/components/submitCode/resetDefaultButton/ResetDefaultButton.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import './resetDefaultButton.css';
-
-function ResetDefaultButton() {
-
-  return (
-    <>
-      <button className='reset-button'>Reset Default</button>
-    </>
-  )
-}
-
-export default ResetDefaultButton
