--- frontend/src/components/submitCode/runButton/RunButton.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import './runButton.css';
-
-function RunButton() {
-
-  return (
-    <>
-      <button className='run-button'>Run</button>
-    </>
-  )
-}
-
-export default RunButton
