--- frontend/src/components/submitCode/runButton/runButton.css
@@ -1,11 +1,0 @@
-.run-button {
-  background-color: green;
-  border-radius: 10px;
-  padding: 0.5% 3% 0.5% 3%;
-  color: white;
-  border-color: white;
-}
-
-.run-button:hover {
-  background-color: darkgreen;
-}
