--- frontend/src/components/submitCode/submitCodeBar/SubmitCodeBar.css
@@ -1,4 +1,0 @@
-.submitCodeBarContainer {
-  background-color: #527595;
-  padding: 0.5% 3% 0.5% 3%;
-}
